:sparkles: Add dynamic navigationViews generation for task notes
Known error: On floating tasks, which do not have notes, launching one of the dynamic navigationViews crashes the application. This is because the task class is used by both floating and daily tasks – however, the floating class has no notes subclass. In the next commit, the task class for floating and daily tasks will be separated in order to fix the bug.