:boom: Add dailyTask and dailyTaskNote permanence using Core Data, breaking the superTask system
The previous system used a global EnvironmentObject to create lists of noteTask and superTask objects. This has been replaced with a Core Data "NoteTask" class, identical to the "noteTask" struct with a few minor changes (priority is stored as Int16, NSManagedObjects, etc), which is updated with quick, asynchronus requests to the phone's storage infrastructure. This allows created tasks and notes to be kept between uses of the application – a fairly standard, very important feature.

The superTask system needs to be reworked to function within the Core Data system. Previously, storing an array of noteTasks within a superTask was no large hassle. However, Core Data "relationships" are much more complicated. This feature will be revisited in the future, and included in the final release, though will not be in active development until after the release of the first alpha. The current plan is to iron out all bugs within the current Core Data infrastructure, introduce time-based clearing of Daily Tasks, and release the app for alpha testers.