Standardize all uses of Text and Image, move completed tasks to the bottom
All non-unique calls of Text and Image (i.e. standard size, limited customization) have been replaced with custom struct wrappers that return similar – if not the same – View as the previous Text with customization. This reduces the amount of lines needed to create the same graphics.

Additionally, completed tasks now have a new look. They are darker than normal tasks, their text is struck through with a line, and they have been moved from the top to the bottom of the FetchRequest so that uncompleted, high-priority tasks are now at the top. As a result of their change in color, the "addTask" and "addNote" structs have been darkened as well (now UIColor.systemGray3).