Add initial functionality to MacOS menu bar app
You can now view and edit tasks from the MacOS popover – changes to completion, priority, and text are synced with other devices using the same iCloud account via CloudKit.

At the moment, there is still some functionality missing from the MacOS companion that is present in the mobile app. There is no onboarding screen, no settings or about, and you cannot view, edit, or add task notes. The task notes feature will be added in the next commit, with other features following quickly after.

There is also one important glitch/bug in this build – for some as-of-yet-unkown reason, the first time you open the popover after launching the app the textfields do not work. Instead, you must begin editing, dissmiss the popover, and open it again without quitting the app. This will be fixed in an upcoming commit.